feat: add authorization

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -49,19 +50,21 @@ public class AuctionsController : ControllerBase
     return Ok(_mapper.Map<AuctionDto>(auction));
   }
 
+  [Authorize]
   [HttpPost]
   public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto createAuctionDto)
   {
     var auction = _mapper.Map<Auction>(createAuctionDto);
 
-    _context.Auctions.Add(auction);
-    auction.Seller = "test";
+    auction.Seller = User.Identity!.Name!;
 
-    var result = await _context.SaveChangesAsync() > 0;
+    _context.Auctions.Add(auction);
 
     var newAction = _mapper.Map<AuctionDto>(auction);
 
     await _publishEndpoint.Publish(_mapper.Map<AuctionCreated>(newAction));
+
+    var result = await _context.SaveChangesAsync() > 0;
 
     if (!result)
       return BadRequest();
@@ -69,6 +72,7 @@ public class AuctionsController : ControllerBase
     return CreatedAtAction(nameof(GetAuction), new { id = auction.Id }, _mapper.Map<AuctionDto>(auction));
   }
 
+  [Authorize]
   [HttpPut("{id}")]
   public async Task<IActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
   {
@@ -78,6 +82,9 @@ public class AuctionsController : ControllerBase
 
     if (auction == null)
       return NotFound();
+
+    if (auction.Seller != User.Identity!.Name)
+      return Unauthorized();
 
     auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
     auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
@@ -95,6 +102,7 @@ public class AuctionsController : ControllerBase
     return Ok();
   }
 
+  [Authorize]
   [HttpDelete("{id}")]
   public async Task<IActionResult> DeleteAuction(Guid id)
   {
@@ -102,6 +110,9 @@ public class AuctionsController : ControllerBase
 
     if (auction == null)
       return NotFound();
+
+    if (auction.Seller != User.Identity!.Name)
+      return Unauthorized();
 
     _context.Auctions.Remove(auction);
